+++ components/Table/components/TableRow/index.tsx
@@ -1,34 +1,28 @@
 import { body1Regular, body2Regular } from "@/styles/fontsStyle";
 import { css } from "@emotion/react";
-import { Data } from "../..";
 import styled from "@emotion/styled";
-import Button from "@/components/Button/Button";
 
-export default function OwnerTableBody({
+type TableData = {
+  name: string;
+  date: string;
+  hourlyPay: string;
+  status: string;
+};
+
+export default function TableBody({
   name,
-  description,
-  phoneNumber,
+  date,
+  hourlyPay,
   status,
-}: Data) {
+}: TableData) {
   return (
     <tbody>
       <TableRow>
         <Cell>{name}</Cell>
-        <Cell>{description}</Cell>
-        <Cell>{phoneNumber}</Cell>
+        <Cell>{date}</Cell>
+        <Cell>{hourlyPay}</Cell>
         <Cell>
-          {status === "대기중" ? (
-            <ButtonContainer>
-              <ButtonWrapper>
-                <Button text="승인하기" color="white" />
-              </ButtonWrapper>
-              <ButtonWrapper>
-                <Button text="거절하기" color="white" />
-              </ButtonWrapper>
-            </ButtonContainer>
-          ) : (
-            <Status status={status}>{status}</Status>
-          )}
+          <Status status={status}>{status}</Status>
         </Cell>
       </TableRow>
     </tbody>
@@ -42,6 +36,11 @@ const getStatusStyle = (status: string) => {
         background: var(--The-julge-blue-10);
         color: var(--The-julge-blue-20);
       `;
+    case "대기중":
+      return css`
+        background: var(--The-julge-green-10);
+        color: var(--The-julge-green-20);
+      `;
     case "거절":
       return css`
         background: var(--The-julge-red);
@@ -51,25 +50,13 @@ const getStatusStyle = (status: string) => {
 };
 
 const TableRow = styled.tr`
+  width: 100%;
   border-bottom: 1px solid var(--The-julge-gray-20);
 `;
 
-const Cell = styled.td`
-  max-width: 300px;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  background: var(--The-julge-gray-05);
+const Cell = styled.td<{ width?: number }>`
   padding: 20px 12px;
   ${body1Regular}
-`;
-
-const ButtonContainer = styled.div`
-  display: flex;
-  gap: 12px;
-`;
-
-const ButtonWrapper = styled.div`
-  width: 92px;
 `;
 
 const Status = styled.div<{ status: string }>`
