+++ pages/my-profile/ApplyHistory/index.tsx
@@ -1,7 +1,8 @@
-import { h1Regular, h3 } from "@/styles/fontsStyle";
-import { Data } from "@/components/Table";
+import { body1Regular, body2Regular, h1Regular, h3 } from "@/styles/fontsStyle";
+import Table, { Data } from "@/components/Table";
 import styled from "@emotion/styled";
-import HistoryTable from "./HistoryTable";
+import Link from "next/link";
+import Button from "@/components/Button/Button";
 
 export type ApplyHistoryProps = {
   type: "employer" | "employee";
@@ -19,18 +20,22 @@ export default function ApplyHistory({
   return (
     <Wrapper>
       <Title>신청 내역</Title>
-      <HistoryTable type={type} limit={limit} count={count} histories={items} />
+
+      {items.length > 0 ? (
+        <Table type={type} limit={limit} count={count} dataList={items} />
+      ) : (
+        <NoApplication>
+          <Description>아직 신청 내역이 없어요</Description>
+          <ButtonContainer>
+            <Link href="/">
+              <Button text="공고 보러가기" />
+            </Link>
+          </ButtonContainer>
+        </NoApplication>
+      )}
     </Wrapper>
   );
 }
-
-const Title = styled.span`
-  ${h1Regular}
-
-  @media (max-width: 767px) {
-    ${h3}
-  }
-`;
 
 const Wrapper = styled.div`
   max-width: 964px;
@@ -49,4 +54,38 @@ const Wrapper = styled.div`
   @media (max-width: 767px) {
     padding: 40px 12px 80px;
   }
+`;
+
+const Title = styled.span`
+  ${h1Regular}
+
+  @media (max-width: 767px) {
+    ${h3}
+  }
+`;
+
+const NoApplication = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  padding: 60px 24px;
+  gap: 24px;
+
+  @media (max-width: 767px) {
+    gap: 16px;
+  }
+`;
+
+const Description = styled.span`
+  ${body1Regular}
+
+  @media (max-width: 767px) {
+    ${body2Regular}
+  }
+`;
+
+const ButtonContainer = styled.div`
+  width: 346px;
+  height: 47px;
 `;
